rename modals from User to Member
