+++ src/groups/GroupList.tsx
@@ -3,13 +3,13 @@ import _ from 'lodash/fp';
 import React, { useEffect, useState } from 'react';
 import { DeleteConfirmationModal } from 'src/components/common';
 import FooterWrapper from 'src/components/FooterWrapper';
-import { Sort } from 'src/components/group-common';
 import { DelayedSearchInput } from 'src/components/input';
 import LeaveResourceModal from 'src/components/LeaveResourceModal';
 import { PageBox, PageBoxVariants } from 'src/components/PageBox';
 import { TopBar } from 'src/components/TopBar';
 import { GroupCard } from 'src/groups/GroupCard';
 import { GroupCardHeaders } from 'src/groups/GroupCardHeaders';
+import { Sort } from 'src/groups/Members/MemberTable';
 import { NewGroupCard } from 'src/groups/NewGroupCard';
 import { NewGroupModal } from 'src/groups/NewGroupModal';
 import { NoGroupsMessage } from 'src/groups/NoGroupsMessage';
